+++ src/mkvtoolnix-gui/util/settings.h
@@ -181,6 +181,7 @@ public:
   unsigned int m_minimumPlaylistDuration;
 
   JobRemovalPolicy m_jobRemovalPolicy, m_jobRemovalOnExitPolicy;
+  unsigned int m_maximumConcurrentJobs;
   bool m_removeOldJobs;
   int m_removeOldJobsDays;
   bool m_useDefaultJobDescription, m_showOutputOfAllJobs, m_switchToJobOutputAfterStarting, m_resetJobWarningErrorCountersOnExit;